backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Google Cloud
google-cloud-aiplatform==1.38.1
google-cloud-speech==2.24.0
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0

# Google Drive Integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# AI/ML
langchain==0.1.0
langchain-google-vertexai==0.0.1

# Data Processing
python-multipart==0.0.6
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0

